requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
